Make aircraft image optional and remove API limits
The image field in the aircraft schema is no longer required, and it has been updated accordingly in the generated types file. Furthermore, the default and maximum limits for the API in the configuration file have been removed to allow for more flexibility in the amount of data returned by API calls.